+++ src/feature/loader/feature-load.spec.ts
@@ -11,8 +11,7 @@ import { ComponentDef } from '../../component';
 import { FeatureContext } from '../feature-context';
 import { FeatureDef } from '../feature-def';
 import { FeatureNeedsError } from '../feature-needs-error';
-import { FeatureKey } from './feature-key.impl';
-import { FeatureLoader } from './feature-loader.impl';
+import { FeatureKey, FeatureLoader } from './feature-loader.impl';
 import { FeatureRequest } from './feature-request.impl';
 import { FeatureRequester } from './feature-requester.impl';
 import Mocked = jest.Mocked;
