+++ src/feature/loader/index.ts
@@ -1,5 +1,4 @@
 export * from './feature-handle.impl';
-export * from './feature-key.impl';
 export * from './feature-request.impl';
 export * from './feature-loader.impl';
 export * from './feature-registry.impl';
